backend/ai_server/requirements.txt
faiss-cpu==1.8.0
sentence-transformers==3.2.1
pypdf==4.2.0
python-docx==1.1.2
pandas==2.2.3
numpy==1.26.4
fastapi==0.115.2
uvicorn==0.32.0
google-generativeai
python-dotenv
